refactor: toss-api.types.ts 를 external/toss 로 이동
- auth.types.ts : UserInfo 를 별도로 선언
