+++ server-toss/src/modules/auth/port/auth-client.interface.ts
@@ -1,8 +1,8 @@
 import { LoginDto } from "../dto/login.dto";
-import { TossUserInfo } from "../types/toss-api.types";
+import { UserInfo } from "../types/auth.types";
 
 export abstract class AuthClient {
   abstract generateToken(dto: LoginDto): Promise<string>;
   abstract removeAccessByUserKey(userKey: number): Promise<void>;
-  abstract getUserInfo(accessToken: string): Promise<TossUserInfo>;
+  abstract getUserInfo(accessToken: string): Promise<UserInfo>;
 }
