+++ server-toss/src/modules/auth/types/auth.types.ts
@@ -1,0 +1,3 @@
+import { TossUserResponse } from "src/external/toss/common/toss-api.types";
+
+export type UserInfo = NonNullable<TossUserResponse["success"]>;
